backend-simple/requirements.txt
fastapi==0.105.0
uvicorn[standard]==0.24.0
openai==1.3.8
python-dotenv==1.0.0
aiofiles==23.2.1
tiktoken==0.5.2
requests==2.31.0
beautifulsoup4==4.12.2
pydantic==2.5.2
python-multipart==0.0.6 